Set default period if it doesn't exist in dependenciesConfig (#2291)
* Set default period if it doesn't exist in dependenciesConfig

This fixes the issue of having to keep the "startTime" value up to date in the template
instance by setting it automatically to one week from the time the dependencyId was first
created in the ledger (i.e. when the user first added the dependency). The default weight
is set to 0.05, or 5% of the total cred.

Test Plan: Create a dependencyConfig that has an empty array for periods, run sourcecred
commands to generate the instance, and ensure the dependencyConfig.json is updated with a
period startTime of one week from the current time.

* Add autoInjectStartingPeriodWeight field to dependency config

Makes things less magical by allowing users to opt into auto-injection of a start period with a default weight

TestPlan:
1. Create a fresh SourceCred instance with autoInjectStartingPeriodWeight unset and ensure that
   the dependency config doesnt have any periods set
2. Create a fresh SourceCred instance with autoInjectStartingPeriodWeight set to 0.05 and ensure that
      the dependency config has a period with the current time and a weight of 0.05

* Change weight validation for dependenciesConfig to non-negative

Having a weight greater than 1 is technically fine. if there's a single dep with weight 1, it gets half
the total cred. if the weight were 2, it would get 2/3rds of the total cred.