Implement MVP of Ledger History Viewer (#2335)
Super basic start for having the ledger event history visible on the frontend. Currently
shows the event name, date / time, and info about the identities for "CREATE_IDENTITY" and
"TOGGLE_ACTIVATION" events. For advanced users, to get the identityId for a specific user/event,
you can hover over the name of the identity and it will appear in a tooltip.

Test Plan:
Ensure the ledger UI is accessible from the "Ledger History" button in the sidebar,
and that it displays a chronological list of events from the ledger with correct dates.