Improve Ledger Admin UX (#2334)
* Rename "Ledger Admin" button to "Identities" in sidebar and change Icon

The ledger admin page is more-so just an identity editor. Grain transfer page also modifies the ledger,
which means "ledger admin" tasks happen across multiple pages, and therefore no single page should be
called "ledger admin". The title on the page itself was also already "Identities"

Test Plan: Check if sidebar button is updated.

* Cleanup styles in AliasView and IdentityMerger

Hardcoded margins are an anti-pattern in material-UI, and in the case of these two
 components aren't necessary

Test Plan: Ensure components still render correctly

* Improve UX of LedgerAdmin page

The editor being at the bottom of the Identity list makes it difficult to edit identities since
users have to scroll through the list, click an identity, and then scroll all the way down to edit it,
all without any visual indicator that a new editor menu is open.

This cleans up the UX by making the list of identities a scrollable box within the page and moving
the editor fields / button to the top of the page so that they stay in the same location and user
isn't required to scroll through potentially thousands of list items. It also improves some of the
labels and buttons to be more descriptive / cleaner.

Test Plan: Ensure there are no UI regressions on the LedgerAdmin page.