Memoize ledger identity calcs in render loop
the ledger identity list is is recalculated on every render. It only
needs to be recalculated if an identity is added or if identities are
merged. This should enhance performance with larger communities where
the list of identites is longer

test plan: load the prod SC cred instance. Update an existing identity
name and ensure the UI keeps up with your typing.