Defer Identity merge to second re-render
Identity merging is a synchronus task. For instances with larger
quantities of participants, this can cause the UI to momentarily freeze
for 1-2 seconds. This effect has been mitigated for the user with two
changes.

1. The Leger merge call was moved to a useEffect hook. This allows
the re-render immediately upon clicking on a target identity.

2. This invokes a loading state, which shows a circular progress wheel
in the right side of the alias text box. This wheel disappears after the
mergeIdentities call finishes.

In production I've noticed more of a hitch than a downright freeze. This
should successfully mitigate it.

test plan: in the production sourceCred instance, merge two identities
that are both composed of more than two ids (decentralion and wchargin are
good go-tos). Ensure there is no freeze, and if the merge take a beat,
you might see the circular progress wheel.