MPG: Expose node and edge order
This commit updates the MarkovProcessGraph so that it exposes a
canonical node and edge order. We further validate that the nodes() and
edges() methods always iterate over nodes and edges in the canonical
order.

This is nice because it makes it cleaner to apply future refactors which
"virtualize the gadgets", i.e. stop representing each gadget node and
edge directly in maps (which is wasteful of memory/disk). Instead, we
can just make sure that the virtualized gadgets are included in the
order, and that the node/edge accessors can retrieve virtualized gadget
nodes and edges.

In a change, we no longer sort the node order. We just want it to be
consistent within a given MarkovProcessGraph and CredGraph, however
other than that nothing is asserted about the order.

Test plan: Some unit tests added, `yarn test` passes, and I also re-ran
CredRank on the real SC instance and verified the scores haven't
changed. Also the CredGraph has a new construction-time invariant
checker which ensures the CredGraph order matches the MarkovProcessGraph
order.