Implement TaskManager (#2266)
TaskManager organizes and maintains a hierarchy of active tasks

It utilizes TaskReporters internally for logging, and allows them to
maintain their own task state for internal use. The primary concern of
the manager is creating and enforcing task scopes via a tree structure.
When a plugin task needs to be terminated in the load CLI command, it is important
that all child tasks also are terminated in case a restart of the task is
necessary. A well-behaved plugin should terminate all tasks it spawns on its own,
but in the case of a cache error or other unexpected error, this class makes it
possible to quickly find and terminate all child tasks.

test plan: unit tests are provided

Co-authored-by: Dandelion Mané <decentralion@dandelion.io>