Eth Plugin: Create the EthAddress file and subcomponents
EthAddress files file stores of ethereum addresses. For now, these
addresses are intended to just be an identity alias for participants
primarily for the purpose of authenticating in a self-service frontend.

Serialized compat files don't preserve the readability of jsonLogs, so
there is little use in utilizing it now. Some compat serialization
logic is needed to preserve readability. This will be important to
implement before wider release in order to better accomodate readable
diffs in a UI and within Git.

test plan: unit tests are provided for the functions utilized by
parsers.