Create the EthAddress file and the needed subcomponents
EthAddress files file stores of ethereum addresses. For now, these
addresses are intended to just be an identity alias for participants
primarily for the purpose of authenticating in a self-service frontend.

This file structure includes a JsonLog with a version header block to ensure
we can upgrade these files as more features are needed in this plugin.

test plan: unit tests are provided for the functions utilized by
parsers. some basic tests of a jsonLog containing EthAddresses might
still be needed if we don't trust the static parse syntax