Virtualize the seed node
This starts the work of virtualizing nodes in the MarkovProcessGraph,
beginning with the seed node. This has a negligible impact on the output
graph size (we only remove one node), but this lets us focus instead on
the new mechanics for supporting virtualization.

Test plan: Tests pass, credrank results are stable, and using the
following `jq` expression before and after the change reveals that the
number of nodes has decreased by exactly one:

```
jq <../cred/output/credGraph.json '.[1].mpg | .[1].nodes | length'
```