credrank: refactor how gadgets are specified (#2368)
This adds a new module, nodeGadgets.js, which provides a consistent
interface for handling node gadgets. This will be quite useful when we
start "virtualizing" the gadget nodes. Also, this reduces the bloat of
the markovProcessGraph module, and allows a separation of concerns
between the core logic of markovProcessGraph, and the incidental logic
of having representation strategies for gadget addresses and nodes.

Test plan: tests pass, also credrank results are stable.