credrank: virtualize participant epochs (#2386)
This follows on rather cleanly from the previous commits which
established the gadgets and the node virtualization strategy. (Really
pleased with my implementation strategy.)

This is the last step of node virtualization and the most significant.
For the first time, we see a material space saving (about 12%). Also,
the number of nodes falls by a factor of 3!

Summary of results so far:

| Step | Virtualize | # Nodes | # Edges | Disk Space |
| ---- | ---------- | ------- | ------- | ---------- |
| 0 | None | 96144 | 416365 | 116M |
| 1 | Seed | 96143 | 416365 | 116M |
| 2 | Accumulators | 96001 | 416365 | 116M |
| 3 | Epochs | 30823 | 416365 | 102M |