credrank: virtualize seed mint edges
Now we virtualize all the seed mint edges. Since we only have mint edges
for nodes with positive weight, this isn't nearly as big a win as
virtualizing all of the radiation edges. But it does still squeeze a few
megs off the output size.

Test plan: Credrank results stable, and tests pass.

| Step | Virtualize | # Nodes | # Edges | Disk Space | % of full |
| ---- | ---------- | ------- | ------- | ---------- | -------- |
| 0 | None | 96144 | 416365 | 116M | 100% |
| 1 | Seed | 96143 | 416365 | 116M | 100% |
| 2 | Accumulators | 96001 | 416365 | 116M | 100% |
| 3 | Epochs | 30823 | 416365 | 102M | 88% |
| 4 | Fibrations | 30823 | 220833 | 66M | 57% |
| 5 | Radiation | 30823 | 124690 | 45M | 39% |
| 6 | Minting | 30823 | 112046 | 42M | 36% |