Revert from JsonLog to array for now
Serialized compat files don't preserve the readability of jsonLogs, so
there is little use in continuing to use it. Some compat serialization
logic is needed to preserve readability. This will be important to
implement before wider release in order to better accomodate readable
diffs in a UI and within Git.

test plan: In the snapshot repo run `sourcecred load` and sourcecred
graph` to ensure the identities are added to the ledger correctly. These
changes are reflected in the changes to the test instance snapshot.