dependency policies include ids
This is a change to dependency policies so that they include the ids as
well as the addresses. This is consistent with the trend in credrank of
tracking participant nodes by ids rather than addresses where possible.

In the future we could possibly remove the address field, but for now
I'm just adding the ids so existing code can be unchanged.

Test plan: `yarn test` passes; also added a new unit test.