discord: cleanup createGraph arguments (#2433)
This commit does a minor refactor of the Discord createGraph method so
that it no longer takes the declaration weights. It's unnecessary for
the createGraph method to know about the default weights, we can mix
them in outside the createGraph method instead. This is more consistent
with the other plugins (e.g. GitHub and Discourse both do it this way),
and lets us remove complexity from the createGraph method, which is nice
because that function is starting to get a bit too complex.

Test plan: It's a trivial change, I checked score parity before/after
making it.