logic for Cred budgets on Weighted Graphs
This commit adds core logic for applying a CredBudget to a
WeightedGraph. The idea is the budget sets constraints on how much Cred
nodes matching a given address can mint each interval. If the budget
constraint is violated in a particular period, all of the nodes created
in that period will be re-weighted to enforce the budget.

Test plan: Thorough unit tests have been added.