api: fix Observable compatibility (#2439)
Summary:
This patch replaces our `getRandomValues` polyfill with a fork that
works in the browser, in Node, in Jest, and in Observable. It’s slightly
hacky because we can’t figure out how to get Webpack’s externs to play
nicely, but the hackiness is entirely contained in this file.

Paired with @topocount.

Fixes #2351.

Test Plan:
Run `yarn build:api`. Test in Node:

```
$ node
> (new (require("./dist/api").default.ledger.ledger.Ledger)()).createIdentity("USER", "z")
'GtJslqp9kl7J8Dr869xGKw'
```

Test in Observable:

 1. Visit <https://observablehq.com/new> and make any change to the
    notebook contents.
 2. Click the “Notebook actions” horizontal ellipsis button to the left
    of the “Publish” button in the top of the screen.
 3. Select “File attachments”.
 4. Click “Upload”, and select `dist/api.js` under your repo.
 5. This will upload `api.js`, or `api@N.js` for some value of `N` if
    you’ve done this before. If you see `api@N.js`, adjust accordingly
    in the following steps.
 6. In an Observable cell, run:

    ```javascript
    {
      const sc = (await require(await FileAttachment("api.js").url()));
      return new sc.default.ledger.ledger.Ledger().createIdentity("USER", "z");
    }
    ```

 7. Observe that it spits out a UUID without crashing.

wchargin-branch: api-observable-build