Pin commonmark to 0.29.1
Commonmark 0.29.2 is a breaking change for us since it scopes the exports
and strictly privatizes the contents of its lib/common and they cannot
be exported.

pinning this dep is necessary because when instance administrators
wipe their yarn.lock files for any reason and re-install, they build
with the 0.29.2 release, which breaks some things from there end.

I understand that this is technically user error, but it's still
something of a Quality-of-Life concern.

test plan: N/A