rename CORE_NODE_PREFIX to GADGET_NODE_PREFIX
The CredRank markov process graph has a concept of "gadgets", which are
nodes that are added specifically for the logic of CredRank. These nodes
are "virtual", in that they are not serialized in the underlying graph.

Heretofore, all of these gadgets have been nested in the
`sourcecred/core` address scope, and we've written an assumption that
every node in that scope is a gadget. However, this will cease holding
as we start adding other nodes in the core scope (e.g. dependency cred
minting nodes). Therefore, we should refine the gadget scope to
`sourcecred/core/gadget`.

Test plan: Existing tests pass, also code inspection. The node gadgets
were already organized around a common prefix (exported as variable)
which I modified and renamed. For the edge gadgets, I added that common
prefix here, and `git grep EdgeAddress.fromParts` in `edgeGadgets.js`
shows that no addresses are being constructed raw (other than the
prefix). Other addresses are all constructed by appending to the prefix.