rename `dependencyMintPolicy` to `bonusMinting`
This commit renames the dependencyMintPolicy module to bonusMinting, and
changes typenames and downstream usage accordingly. This is motivated,
most proximately, by my need to re-write this system to output a bonus
minting subgraph as part of the CredRank integration. (We want to have
dependency minting occur directly within the Graph, rather than adding
extra logic to the whole score computation pipeline.)

While writing the graph module, I found that the naming became a big
mess. The problem, I think, is that "DependencyMint" is a confusing
name: it describes the intended use case (paying dependencies) but not
the mechanism at hand (minting a bonus %age of total Cred that flows to
a particular recipient). I think renaming the mechanism to better
reflect its behavior will make the core logic more readable.

I've left the config API un-renamed, i.e. the config file is still
called dependnecies.json. That makes sense, since the intended use case
is still for minting Cred for dependencies.

Test plan: All existing tests pass. It's just a bunch of renames so that
suffices.