grainConfig: implement Recent configuration
This adds configuration logic for the Recent policy to read from an instance's `config/grain.json` file.  Here, we add an optional `discount` parameter to the `GrainConfig` type, which checked for if there is a non-zero Recent policy.

This is followup to #2447

__Test Plan__
I've added some unit tests covering the `GrainConfig` parser.  I also added tests to cover errors related to missing a missing `discount`.