Bugfix: create output directory in score command
In the template instance, no prior commands yield anything in the output
directory. This yields a bug in the score command because there is no
allowance for it not to exist at this point. Adding a mkdirx command
here fixes that.

test plan: run `scdev go` in a clean template instance and verify the
folder is created without and filesystem-level errors throwing