Refactor computeCredAccounts to function in the empty case
The public computeCredAccounts command would fail if no intervals had
been computed yet, because userNodes were always assumed to contain some
cred values. This refactoring allows for computeCredAccounts to succeed
without any credIntervals needing to be created.

This is needed to allow the template instance to successfully run `yarn
go` out of the box (even if it doesn't do anything substantial without
    plugins)

test plan: The outer function is untested, but the inner function has
been updated to demonstrate support for both the empty case and the
"default case", meaning the case in the template instance