Discord: Refactor createGraph (#2468)
This refactors the createGraph function in the Discord plugin to separate
data processing from graph creation. Basically, it splits it into a
function called `_findGraphMessages` which yields all of the information
needed to produce the graph, and `_createGraphFromMessages` generates
the actual graph using this data.

Doing it this way allows us to simplify the control flow. As an example:
we have a rule that we don't emit message nodes for messages with no
reactions or mentions. Previously, this was tracked via a `hasEdges`
boolean which would track whether we had added any reaction or mention
edges. Now, the `findGraphMessages` function explicitly only yields
messages that have mentions or reactions.

This refactor makes it easier to write unit tests (and allows us to
prioritize testing the areas that will have more logic and control flow
cases, i.e. the `_findGraphMessages` function). However, I haven't
actually added such tests, mostly because producing a mock
`SqliteMirrorRepository` for testing purposes will be a pain, and I'm
short on bandwidth. I'm comfortable merging this sans-testing because
I think it's a meaningful improvement/cleanup of the current logic
(which is also untested).

Co-authored-by: Dandelion Mané <decentralion@dandelion.io>