Bugfix: support empty JsonLog files (#2486)
We support legacy empty-case formats, but we don't support
the empty-case for ndJson. This change fixes that, by adding support for
when the serialized logfile contains no entries.

This is technically a bug because when an empty JSONlog is serialized it
returns an empty file and there was no way to parse it again. This case
presented itself when running `yarn graph` and `yarn score` in a clean
template instance

test plan: unit tests have been updated to demonstrate support for this
case.