credrank: provide cleaner interface
This refactors the interface in `credrank/compute.js`. Now, the caller
provides a `WeightedGraph` and a `Ledger`, instead of providing a
`MarkovProcessGraph`. This is appropriate because the
`MarkovProcessGraph` is an internal concept to the credrank
implementation, whereas the `WeightedGraph` and `Ledger` are
broadly-available concepts.

This is immediately motivated by wanting to write tests for
`credgraph.js` without needing to manually construct a
`MarkovProcessGraph`. Producing a `WeightedGraph` and `Ledger` so as to
construct a test case will be cleaner. Also, this allows a cleanup of
the `cli/credrank` command.

Test plan:
The `credrank/compute` module is untested, because it is the kind of
data-pipelining module that tends to be annoying to unit test, but is
also very unlikely to be broken in ways that flow doesn't catch. As a
precaution, I ran the credrank command and verified that scores are
still as expected.