grainConfig: take array of policies
This commit allows users to specify an array of policies, which allows them two new advantages:
1. Do not need to specify each policy.
2. Can add multiple version of the same policy (in the case that you want to use different parameters).

Currently, users specify a budget for every valid policy in their instance's `config/grain.json` file, even if they are not using that grain allocation policy.
```js
{
  "immediatePerWeek": 0,
  "balancedPerWeek": 10,
  "recentPerWeek": 0,
  "maxSimultaneousDistributions": 100
}
```

This also forces us to handle each policy explicitly downstream, which could be a lot cleaner if we could leverage the `AllocationPolicy` type we already have.  This approach uses typing to check whether the user has specified a valid policy implicitly.

Config now looks (for example) as follows:
```js
{
  policies: [
    {
      policyType: "BALANCED",
      budget: "100"
    },
    {
      policyType: "RECENT",
      budget: "50",
      discount: 0.1
    }
}
```

__Test Plan__
Unit tests are provided here for both the `parser`, and well as `toDistributionPolicy`, which serves to check the config.  Note, we can reduce this function entirely with stricter parsers and types.  For example, we have to check for negative grain values.  This calls for a new "`GrainBudget"` type, which I will follow up with.