Keep the user from losing their place on the ledger view when changing # of results per page (#2507)
In reviewing the changes for the recent update to the Ledger Viewer page from PR #2389 I noticed that when changing the number of results to show per page, the user would automatically be reset to the first page. For a more seamless user experience, I've made some small changes so that the user continues to be in (at least approximately) the same place in the result set, even after changing the number of results per page. Because the resulting page size may be larger or smaller than the previous size, I defaulting to ensuring the new page number includes the first row of the results the user was viewing (as opposed to, say, the middle or final row). The intention of this more conservative choice is to make sure the user is never moved further down in the result set, thereby unintentionally skipping results.

Test Plan
I launched my local template-instance and tried every combination of result size changes (25->50, 50->100, 100->25, etc). All resulted in the expect change of remaining on the page number that contained the first row of the previous result page.