grainConfig: optional budgets (option 2)
This commit adds support for omitted budgets in a SourceCred instance's `config/grain.json` file.

Here, the `GrainConfig` type takes the policy budgets optionally, and passes the responsibilty to handle null values to `toDistributionPolicy`.  This is an alternative implementation to #2522.  The advantage here is that we consolidate all the configuration checks on the data in `toDistributionPolicy`, as opposed to splitting its responsibilities with the parser.

__Test Plan__
I've modified the parser's unit tests to allow missing budget fields.  In addition, I've modified unit tests for `toDistributionPolicy` to test that missing parameters are handled properly.