Discard all local changes in LedgerManager if there's any conflicts
Instead of keeping the successfully applied local changes, this changes it to discard everything
and reset to the remote state to ensure atomic transactions when writing to the ledger. This comes
with the downside that it might be tedious for the user to re-apply all the changes they made, but
the front end UX can help alleviate this issue if we regularly sync the remote ledger.

Test Plan: unit tests pass