Implement LedgerManager class (#2480)
* Implement LedgerManager class

The LedgerManager will handle the diffing and syncing of Ledger event logs between an in-memory
ledger and a "remote ledger" through a generic LedgerStorage interface, allowing it to be persisted
to a variety of storage backends (e.g. GitHub, filesystem, database, etc).

Test Plan:
- Unit tests pass
- Interface makes sense for initial use case of GitHub storage

* Clarify resetting remote state in LedgerManager and add tests

Updated comment to clarify that the successfully applied local changes arent reverted when
a conflicted change fails to be applied and added a test case to verify this functionality.

Test Plan: Unit tests pass

* Discard all local changes in LedgerManager if there's any conflicts

Instead of keeping the successfully applied local changes, this changes it to discard everything
and reset to the remote state to ensure atomic transactions when writing to the ledger. This comes
with the downside that it might be tedious for the user to re-apply all the changes they made, but
the front end UX can help alleviate this issue if we regularly sync the remote ledger.

Test Plan: unit tests pass