immediate policy: add lookback parameter
Right now, the immediate policy uses cred from the most recent interval.  This generalizes
it such that users can create policies based on the past N weeks, without the discounting
of the Recent policy.

This includes an optional `numInteravlsLookback` parameter in the `ImmediatePolicy` type, which
defaults to 1 in config until #2600 gets merged, at which point we'll allow users to specify
alternative lookback lengths.

Note, this raises a question around naming, as Recent is actually a better name here, and we
could use something for the current Recent policies that highlights the discounting more.

__Test Plan__
Unit tests provided check that the `immediatePolicy` works for valid params, and also checks
for invalid `numIntervalsLookback` parameters.

This also updates the snapshot in `core/ledger/ledger.test` to include this extra paramter
on the `ImmediatePolicy`.