add support for limiting Cred minting by plugin (#2547)
This commit builds on #2512 to allow instances to limit the rate of Cred
minting on a per-plugin basis. For example, an instance maintainer could
set a rule where the GitHub plugin can only mint up to 100 Cred per
week. The budgets may also vary over time; for example, the budget could
be 100 Cred per week in the year 2019, but 200 Cred per week in 2020.

This change is motivated in part by 1Hive's experience using SourceCred,
where the activity level on the Discord plugin grew much faster than
activity on other plugins, and threatened to drown out other Cred
minting.

To activate this feature, simply add a `config/pluginBudgets.json` file.
Here's an example:

```json
{
  "intervalLength": "WEEKLY",
  "plugins": {
    "sourcecred/github": [
      {"startTime": "2018-01-01", "budget": 10}
    ],
    "sourcecred/discord": [
      {"startTime": "2018-01-01", "budget": 10}
    ],
    "sourcecred/discourse": [
      {"startTime": "2018-01-01", "budget": 10}
    ],
    "sourcecred/initiatives": [
      {"startTime": "2018-01-01", "budget": 10}
    ]
  }
}
```

Note that the budget limitation is a _maximum_, i.e. a plugin might
still mint less than the stated amount in a given period. This means
that it's not possible to write strict invariants like "GitHub mints
half and Discord mints half". If there's a week with very little GitHub
activity, the GitHub plugin is likely to come in under-budget, in which
case the GitHub plugin will mint less than half of the total Cred.

Test plan: The code does not have any unit tests. This is typical of
"plumbing" type code where the logic basically amounts to deserializing
a config file and then piping it into the core logic: it's hard to get
things "subtly wrong" where it breaks in surprising ways. Flow passes,
which suggests things are wired correclty. I've also manually tested
this by imposing a variety of budget constraints on the real SourceCred
instance, and can confirm that things work as intended. For example,
here's the `credrank` output when applying a limitation on the SC
instance that each plugin may mint at most 10 Cred per week:

| Description | Cred | % |
| --- | --- | --- |
| decentralion | 1111.2 | 34.0% |
| wchargin | 413.5 | 12.6% |
| beanow | 257.2 | 7.9% |
| s-ben | 214.3 | 6.6% |
| lbstrobbe | 141.4 | 4.3% |
| mzargham | 107.4 | 3.3% |
| hammad | 96.1 | 2.9% |
| brianlitwin | 71.3 | 2.2% |
| burrrata | 67.8 | 2.1% |
| topocount | 62.3 | 1.9% |
| bex | 55.2 | 1.7% |
| KuraFire | 50.3 | 1.5% |
| dependabot | 36.3 | 1.1% |
| greenkeeper | 32.2 | 1.0% |
| panchomiguel | 31.6 | 1.0% |
| vsoch | 31.5 | 1.0% |
| sourcecred | 29.3 | 0.9% |
| evan | 29.3 | 0.9% |
| sandpiper | 24.9 | 0.8% |
| benoxmo | 19.3 | 0.6% |