Add tested support for deserializing CredGraphs
This commit re-writes the (de)serialization pathway for CredGraph, which
was previously not functional. I've also rewritten the path to use
parsers.

I removed any usage of the compat module. The main function of the
compat module (especially when considering ephemerally-maintained data
like the CredGraph) is to ensure we get clean error messages if the
provided data is invalid. Happily, parsers do that without requiring any
extra bookkeeping on our part. The compat module makes it slightly
easier to upgrade data formats while preserving backwards compatibility,
but since cred graphs are regenerated every time the `credrank` command
is run, this is not an important consideration.

The new logic is unit tested for both the MarkovProcessGraph and the
CredGraph. I also added a helper function to `cli/common` for retrieving
the CredGraph. This doesn't have explicit unit tests, but I manually
tested it by patching the `cli/credrank` command to load the Cred Graph
from disk before reporting results.