Refactor Credrank Testing & Test CredGraph
This commit sets up testing for the CredGraph module, which was
previously entirely untested. Getting an example CredGraph is a bit of a
pain, because it requires constructing a MarkovProcessGraph, which is
itself fairly involved.

We already had logic for constructing an example MarkovProcessGraph, for
MPG's own testing, and I didn't want to duplicate it. Instead, I
factored out all the example logic from `markovProcessGraph.test.js`
into `credrank/testUtils.js`. This is similar to the pattern we have for
testing regular Graphs, see `core/graphTestUtil.js`.

I then set up some basic sanity-check testing in `credGraph.test.js`,
mostly to show that it's now possible.

Test plan: This is a test-only change. `yarn test` passes.