grain cli: add simulation flag (#2545)
This adds an optional `--sim` flag which allows users to run a simulation of the upcoming grain distribution without writing to the ledger or accounts.  This will be more useful as we allow more advanced simulation configurations, and should not be considered final.

Ultimately, we will want to create simulated ledger and accounts data which can be analyzed more easily than copying and pasting a markdown table.  This also could serve as the foundation for simulating multiple distributions that a community could vote on.

__Test Plan:__
Trivial change and sanity checked against the similar code in `src/cli/go.js`.  Tested myself using `yarn grain` and `yarn grain --sim`, checking my local cred instance's ledger for changes.