Remove Recharts Dependency
I went to work on the mobile responsitivity of the sparkline chart in the Explorer in the FE. I assumed it was going to be in Recharts after seeing it listed in the project's dependencies; however it appears D3 is being used for charts. I looked through the repo and it appears Recharts is completely unused. Let's remove it so that it doesn't bloat our build, introduce potential security risks, and just for the sake of Clean Code.

To test out, just run `yarn` and `yarn test` to ensure that removing Recharts has not effected the project.