Support Mobile View on Explorer Page
#Description
Currently, in the explorer view, the D3 sparkline charts break screen width responsitivity. This simple fix allows the SVG chart to scale and thus allows this page to be more responsive.

#Test
Run the site locally (`yarn start`) and go to the Explorer view. (which is also the default route).
Resize your screen or check it out in the chrome inspector and note that scrolling right in the mobile view no longer has the whitespace it did before.