Weights type: uniformly import as WeightT
__Context__
The `Weights` type is currently imported as either `Weights` or `WeightsT`.  The latter case is
used when the type conflicts with the `core/weights.js` module which is imported as `Weights` as
well.

__Advantages__
This renames the `Weights` type to `WeightsT` and changes all imports/instances of the type formerly
known as `Weights`.  This gives us consistency, which has the following advantage:

(1) Searching for instances of the `Weights` (now `WeightsT`) type is much simpler as we don't need
to also sort through instances of importing the `core/weights.js` module (and vice versa).

__Test Plan__
Manually checked import instances of the `core/weights.js` module for imports, and used Flow to check
for the remaining cases within those files.

Can also search for import instances using `git grep "import \(type \)\?{.*Weights.*}"`.