CredRank: Remove infinite "padding" intervals
When we create a MarkovProcessGraph, partition the history of the Graph
into "time intervals" and attribute participant Cred within each
interval. At present, we also add two extra infinite intervals: one
covering all time before the start of the first regular interval (i.e.
the range `[-Infinity, firstIntervalStart)`), and one covering all time
after the end of the last regular interval (i.e. the range
`[lastIntervalEnd, +Infinity]`). We did so in order to avoid Cred
distortions at the boundary, where the first and last intervals get
Cred scores that are biased upwards or downwards due to their position
in the ordering.

However, this adds a fair bit of complexity. It means that we can
no longer directly correspond Cred scores onto consistent,
finite-duration intervals: we need to expose this concept of padding
intervals to all consumers. (If we try to hide the existence of the
padding intervals, then we come into conflict with other invariants we
care about, like that a participant's Cred is the sum of their Cred
within each interval, and that total participant Cred sums to total
minted Cred).

Based on discussion with @wchargin and @blueridger, I've decided to
take a different approach. We now no longer add extra padding intervals.
In order to avoid Cred distortions, we instead add a "looped" edge,
which connects the first epoch to itself via the "backwards" edge, and
the last epoch to itself via its "forward" edge. This is a very nice
solution as it removes bias that actually existed in the "padding
intervals" approach, because even with padding intervals, there was
still a bit of excess Cred vanishing to the seed (Cred that flowed out
to the padding interval, and should have kept going "forward" or
"backward", but instead reset to seed).

Test plan: Unit tests updated. Credrank scores change, very slightly, as
might be expected. We can't use `credrank -d` because the semantics of
the serialized graph changes. However, raw before/after tables are
below. In general, people with lots of recent activity have seen their
Cred increase by a tiny amount (order of 0.1% relative).

Before:

| Description | Cred | % |
| --- | --- | --- |
| decentralion | 25828.3 | 29.0% |
| wchargin | 10857.9 | 12.2% |
| beanow | 5774.9 | 6.5% |
| lbstrobbe | 5749.9 | 6.5% |
| s-ben | 5715.9 | 6.4% |
| hammad | 4082.2 | 4.6% |
| topocount | 2444.0 | 2.7% |
| KuraFire | 2265.4 | 2.5% |
| burrrata | 1944.5 | 2.2% |
| bex | 1901.7 | 2.1% |
| panchomiguel | 1733.5 | 1.9% |
| sourcecred | 1598.7 | 1.8% |
| sandpiper | 1577.2 | 1.8% |
| dependabot | 1552.7 | 1.7% |
| mzargham | 1432.6 | 1.6% |
| joiecousins | 1076.1 | 1.2% |
| Thena | 1070.1 | 1.2% |
| brianlitwin | 893.4 | 1.0% |
| vsoch | 793.2 | 0.9% |
| greenkeeper | 662.8 | 0.7% |

After:

| Description | Cred | % |
| --- | --- | --- |
| decentralion | 25825.7 | 29.0% |
| wchargin | 10856.2 | 12.2% |
| beanow | 5772.8 | 6.5% |
| lbstrobbe | 5749.3 | 6.5% |
| s-ben | 5714.5 | 6.4% |
| hammad | 4081.7 | 4.6% |
| topocount | 2443.7 | 2.7% |
| KuraFire | 2266.1 | 2.5% |
| burrrata | 1943.7 | 2.2% |
| bex | 1901.4 | 2.1% |
| panchomiguel | 1734.8 | 1.9% |
| sourcecred | 1598.1 | 1.8% |
| sandpiper | 1577.7 | 1.8% |
| dependabot | 1552.4 | 1.7% |
| mzargham | 1432.1 | 1.6% |
| joiecousins | 1077.8 | 1.2% |
| Thena | 1072.6 | 1.2% |
| brianlitwin | 893.1 | 1.0% |
| vsoch | 792.9 | 0.9% |
| greenkeeper | 662.5 | 0.7% |